+++ planetary_test_data/__init__.py
@@ -2,7 +2,9 @@
 
 __author__ = 'PlanetaryPy Developers'
 __email__ = 'contact@planetarypy.com'
-__version__ = '0.3.3'
+__version__ = '0.4.0'
 
 
 from .planetary_test_data import PlanetaryTestDataProducts
+
+__all__ = ['PlanetaryTestDataProducts']